Add files via upload
Initial run on multi-linear regression.  Used an older data set and can update or test newer data sets.  r2 score was pretty good - not sure I trust it.  Need to do a bit more work to see if this is really right .  Also want to test one or two other regression models (Poisson, LARS) to see if they are valid.